Split generic code out of AbstractMetastoreTableOperations
